+++ koin-projects/examples/androidx-samples/src/test/java/CheckModulesTest.kt
@@ -1,24 +1,27 @@
+import android.app.Activity
+import org.junit.Rule
 import org.junit.Test
 import org.junit.experimental.categories.Category
-import org.koin.core.parameter.parametersOf
-import org.koin.sample.androidx.components.mvp.FactoryPresenter
-import org.koin.sample.androidx.components.mvp.ScopedPresenter
-import org.koin.sample.androidx.di.appModule
-import org.koin.sample.androidx.di.mvpModule
+import org.koin.core.logger.Level
+import org.koin.dsl.module
+import org.koin.sample.androidx.di.allModules
 import org.koin.test.category.CheckModuleTest
 import org.koin.test.check.checkModules
+import org.koin.test.mock.MockProviderRule
+import org.mockito.Mockito
 
 @Category(CheckModuleTest::class)
 class CheckModulesTest {
 
+    @get:Rule
+    val mockProvider = MockProviderRule.create { clazz ->
+        Mockito.mock(clazz.java)
+    }
+
     @Test
     fun `test DI modules`() =
-        checkModules(
-            parameters = {
-                create<FactoryPresenter> { parametersOf("_ID_") }
-                create<ScopedPresenter> { parametersOf("_ID_") }
-            }
-        ) {
-            modules(appModule + mvpModule)
+        checkModules {
+            printLogger(Level.DEBUG)
+            modules(allModules)
         }
 }